# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BigFin API
  version: 0.1.0
  description: |
    Multi-tenant loan administration API powering GreenBucks and other operator apps.

    ## Key Concepts
    - **Tenant**: Operator app (e.g., GreenBucks) with isolated data
    - **EndCustomer**: Borrower or lender on the platform
    - **LoanOffer**: Proposed loan terms from lender to borrower
    - **LoanContract**: Accepted offer with repayment schedule
    - **PaymentIntent**: Idempotent payment request (disbursement or repayment)

    ## Idempotency
    All POST endpoints that create resources or initiate payments accept an `Idempotency-Key` header.
    Requests with the same key within 24 hours return the original response.

    ## Error Codes
    See the Error Taxonomy section at the bottom of this spec.
  contact:
    name: BigFin API Support
  license:
    name: Proprietary
servers:
  - url: https://api.bigfin.io/v1
    description: Production
  - url: https://sandbox.api.bigfin.io/v1
    description: Sandbox
security:
  - BearerAuth: []
...
paths:
...
  /loan-contracts/{contract_id}/disburse:
    parameters:
      - $ref: '#/components/parameters/ContractId'
    post:
      tags: [Disbursements]
      summary: Initiate disbursement
      description: |
        Initiates fund transfer to borrower. Speed can be `standard` (ACH) or `instant`
        (RTP/FedNow/push-to-card with fallback).

        If `speed=instant` and lender has sufficient prefund, express fee is waived.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisbursementRequest'
      responses:
        '202':
          description: Disbursement initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disbursement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Loan not in pending_disbursement status
          $ref: '#/components/responses/Conflict'
  /disbursements/{disbursement_id}:
    parameters:
      - $ref: '#/components/parameters/DisbursementId'
    get:
      tags: [Disbursements]
      summary: Get disbursement status
      responses:
        '200':
          description: Disbursement details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disbursement'

  # ============ REPAYMENTS ============
...
  /operator/loans:
    get:
      tags: [Operator]
      summary: Search loans
      description: Operator search across all loans with filters
      parameters:
        - name: q
          in: query
          description: Search query (customer name, email, loan ID)
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_disbursement, active, paid_off, defaulted, cancelled]
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanContractList'
  /operator/ledger-journals/{journal_id}/reverse:
    parameters:
      - name: journal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Operator]
      summary: Reverse ledger journal
      description: |
        Creates a reversing journal entry. Requires step-up authentication.
        Original journal is not modified (append-only ledger).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: X-Step-Up-Token
          in: header
          required: true
          schema:
            type: string
          description: Step-up authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 10
                  maxLength: 1000
                  description: Detailed reason for reversal (audit requirement)
      responses:
        '201':
          description: Reversal journal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerJournal'
        '400':
          description: Journal already reversed
          $ref: '#/components/responses/BadRequest'
  /operator/reconciliation/exceptions:
    get:
      tags: [Operator]
      summary: List reconciliation exceptions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, investigating, resolved]
        - name: date
          in: query
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Exception list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationExceptionList'
  /operator/reconciliation/exceptions/{exception_id}:
    parameters:
      - name: exception_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Operator]
      summary: Update exception status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [investigating, resolved]
                resolution_notes:
                  type: string
      responses:
        '200':
          description: Exception updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationException'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Unique key for idempotent requests. Same key within 24h returns cached response.
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
    PageLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ContractId:
      name: contract_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DisbursementId:
      name: disbursement_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    LoanTerms:
      type: object
      required: [principal_cents, apr_bps, term_months, payment_frequency]
      properties:
        principal_cents:
          type: integer
          minimum: 10000
          maximum: 5000000
        apr_bps:
          type: integer
          minimum: 0
          maximum: 3600
          description: APR in basis points (e.g., 1200 = 12.00%)
        term_months:
          type: integer
          minimum: 1
          maximum: 60
        payment_frequency:
          type: string
          enum: [weekly, biweekly, monthly]
        first_payment_date:
          type: string
          format: date
          description: If not specified, calculated based on disbursement + min days
...
    LoanContract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        offer_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        lender_id:
          type: string
          format: uuid
        borrower_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending_disbursement, active, paid_off, defaulted, cancelled]
        terms:
          $ref: '#/components/schemas/LoanTerms'
        disbursement:
          $ref: '#/components/schemas/DisbursementSummary'
        balances:
          $ref: '#/components/schemas/LoanBalances'
        next_payment:
          $ref: '#/components/schemas/ScheduledPayment'
        originated_at:
          type: string
          format: date-time
        disbursed_at:
          type: string
          format: date-time
        paid_off_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    LoanContractList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LoanContract'
        next_cursor:
          type: string
        has_more:
          type: boolean
    LoanBalances:
      type: object
      properties:
        principal_cents:
          type: integer
        interest_cents:
          type: integer
        fees_cents:
          type: integer
        total_due_cents:
          type: integer
        past_due_cents:
          type: integer
        payoff_amount_cents:
          type: integer
          description: Amount to pay off loan today
...
    ScheduledPayment:
      type: object
      properties:
        sequence:
          type: integer
        due_date:
          type: string
          format: date
        principal_cents:
          type: integer
        interest_cents:
          type: integer
        fees_cents:
          type: integer
        total_cents:
          type: integer
        status:
          type: string
          enum: [scheduled, due, paid, partial, past_due]
        paid_cents:
          type: integer
        paid_at:
          type: string
          format: date-time

    # ===== Disbursements =====
    DisbursementRequest:
      type: object
      required: [speed]
      properties:
        speed:
          type: string
          enum: [standard, instant]
        funding_instrument_id:
          type: string
          format: uuid
          description: Override destination (defaults to contract's instrument)
    Disbursement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
        amount_cents:
          type: integer
        express_fee_cents:
          type: integer
        net_amount_cents:
          type: integer
          description: Amount after fee deduction
        speed:
          type: string
          enum: [standard, instant]
        rail:
          type: string
          enum: [ach, same_day_ach, rtp, fednow, push_to_card]
        status:
          type: string
          enum: [initiated, pending, completed, failed]
        availability:
          $ref: '#/components/schemas/FundsAvailability'
        funding_instrument_id:
          type: string
          format: uuid
        source:
          type: string
          enum: [prefund, direct]
          description: Whether funded from prefund or direct transfer
        initiated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        failure_reason:
          type: string
    DisbursementSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount_cents:
          type: integer
        express_fee_cents:
          type: integer
        net_amount_cents:
          type: integer
        speed:
          type: string
        rail:
          type: string
        status:
          type: string
        completed_at:
          type: string
          format: date-time
    FundsAvailability:
      type: object
      properties:
        state:
          type: string
          enum: [initiated, pending, received, held, available, failed]
        available_at:
          type: string
          format: date-time
        hold_reason:
          type: string
        hold_expires_at:
          type: string
          format: date-time

    # ===== Repayments =====
...
    LedgerJournal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [disbursement, repayment, fee, interest_accrual, reversal, adjustment]
        description:
          type: string
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntry'
        is_reversal:
          type: boolean
        reverses_journal_id:
          type: string
          format: uuid
        reversed_by_journal_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
    LedgerEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        journal_id:
          type: string
          format: uuid
        account:
          type: string
          description: Hierarchical account code (e.g., "assets:loans_receivable:principal")
        debit_cents:
          type: integer
        credit_cents:
          type: integer
        balance_after_cents:
          type: integer
        created_at:
          type: string
          format: date-time
...
    ReconciliationException:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        type:
          type: string
          enum: [balance_mismatch, stuck_payment, duplicate_transaction, missing_webhook, other]
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        affected_entity_type:
          type: string
        affected_entity_id:
          type: string
          format: uuid
        expected_value:
          type: string
        actual_value:
          type: string
        status:
          type: string
          enum: [open, investigating, resolved]
        resolution_notes:
          type: string
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        resolved_by:
          type: string
          format: uuid
    ReconciliationExceptionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReconciliationException'
        next_cursor:
          type: string
        has_more:
          type: boolean

    # ===== Errors =====
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable message
        details:
          type: object
          description: Additional context
        request_id:
          type: string
          description: Request ID for support
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "invalid_request"
            message: "Principal must be between $100 and $50,000"
            request_id: "req_abc123"
...
    Conflict:
      description: Conflict with current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "invalid_state"
            message: "Offer has already been accepted"

# ===== Error Taxonomy =====
#
# Error codes follow the pattern: {category}_{specific}
#
# Authentication/Authorization:
#   - unauthorized: Missing or invalid token
#   - forbidden: Valid token but insufficient permissions
#   - step_up_required: Sensitive action requires re-authentication
#
# Validation:
#   - invalid_request: Malformed request body
#   - invalid_parameter: Specific parameter validation failed
#   - terms_out_of_policy: Loan terms violate product policy
#
# State:
#   - invalid_state: Action not allowed in current state
#   - already_exists: Duplicate resource (use idempotency key)
#   - not_found: Resource doesn't exist
#
# Payment:
#   - insufficient_funds: Not enough balance
#   - instrument_invalid: Funding instrument can't be used
#   - payment_failed: Payment provider rejected
#   - payment_returned: Payment was returned/reversed
#
# Limits:
#   - limit_exceeded: Transaction or balance limit hit
#   - rate_limited: Too many requests
#
# System:
#   - internal_error: Unexpected server error
#   - service_unavailable: Temporary outage
#   - provider_error: Payment provider issue